progress

+++ mkfs.wfs.c
@@ -43,7 +43,7 @@ void initialize_filesystem(const char *disk_path) {
     struct wfs_sb* superblock = (struct wfs_sb*)base;
 
     superblock->magic = WFS_MAGIC;
-    superblock->head = ((int)base) + sizeof(struct wfs_sb);
+    superblock->head = ((uintptr_t)base) + sizeof(struct wfs_sb);
  
     // Initialize the root directory log entry
     struct wfs_inode root_inode;
